Add garbage collection
Previously, I was using shared_ptrs for managing the memory for
instances, but there was a problem where instances could create
cyclical references to each other, and the memory for instances
would never be freed. I initially tried to use a combination of
weak_ptrs and shared_ptrs to solve this problem, but eventually
I just ended up using a simple mark-and-sweep garbage collector
for managing memory. The collector could definitely be improved
but it's good enough for now until someone actually needs decent
performance from their Glass program.