Add support for inheritance
I decided to add an extension to the language to add support for
inheritance, because I'm not really sure why the original language
doesn't support it. I also made it so the compiler can convert code
written with inheritance to standard-conforming glass, and added a
compiler flag to reject code written with extensions, if you want
your code to be portable.