Small improvement to minification
When the same name is pushed to the stack twice in a row, the second
push is changed to duplicate the top element of the stack, so that if
the name is at least two characters long, it actually saves three
characters. This can probably be extended later to perform more
aggressive minification.