Reference count strings in compiled Glass code
So this changes the compiler such that the compiled code uses reference
counting for managing the memory for strings. This doesn't work
perfectly yet, since the memory for instances is not yet managed, so if
a string is referenced by an instance, it will never be freed. It seems
to work so far for strings on the stack and in local and global
variables, so when garbage collection for instances is implemented, it
should work prefectly (fingers crossed).