Add optimization for function calls
In Glass code, I noticed that pretty much all function calls are done
in the same exact way, so I added an optimization stage to collapse the
four-command sequence into a single command. This avoids pushing names
to the stack when unnecessary, plus it makes the compiled C code a
great deal shorter.