Use name mangling to ensure unique compiled names
I realized that it was possible to name functions and classes in Glass
code such that when you compiled it, the names for two different
functions would end up being the same, and thus illegal C code. This
code does some name mangling such that it is impossible for two
different methods to have the same compiled name.