Allow space between name and control structures
A few days ago, I fixed a "bug" where my minification tool could split
up a name with the beginning of a class or function. However, after
doing some testing with the reference interpreter, I discovered that it
was perfectly permissible for the name of a class, function or loop to
be separated from the preceding "{", "[" or "/" character by
whitespace. This commit makes it so that this interpreter accepts such
input, and makes it so the minification tool is allowed to generate
such code.