Remove ability to output names
So the wiki page for Glass claims that the O.o function can output both
strings and names, which is true of the reference interpreter. However,
after looking through the edit history of the wiki and the source to
the reference interpreter, I don't think it was ever intentional for
O.o to output names. I think the only reason O.o works for names in the
reference interpreter is because none of the built-in functions in the
reference interpreter bother checking that things are of the right
type. Since my interpreter is a great deal stricter than the reference
one, I've decided to remove the ability of O.o to output names.