Fix constructor inheritance
Previously, if there was a return command in a parent's constructor, it
would completely quit out of the construction process. This makes it so
instead of copy-and-pasting the code from the parent constructor to the
beginning of the child constructor, it copies the parent's constructor
to a new function, and has the child's constructor call that new
function.